+++ Kernel/Output/HTML/Layout/Loader.pm
@@ -11,6 +11,11 @@ package Kernel::Output::HTML::Layout::Loader;
 use strict;
 use warnings;
 
+use File::stat;
+use Digest::MD5;
+
+use Kernel::Language qw(Translatable);
+
 our $ObjectManagerDisabled = 1;
 
 =head1 NAME
@@ -238,6 +243,206 @@ sub LoaderCreateAgentJSCalls {
         );
 
     }
+
+    return 1;
+}
+
+=item LoaderCreateJavaScriptTemplateData()
+
+Generate a minified file for the template data that
+needs to be present on the client side for JS based templates.
+
+    $LayoutObject->LoaderCreateJavaScriptTemplateData();
+
+=cut
+
+sub LoaderCreateJavaScriptTemplateData {
+    my ( $Self, %Param ) = @_;
+
+    # get config object
+    my $ConfigObject = $Kernel::OM->Get('Kernel::Config');
+
+    # load theme
+    my $Theme = $Self->{UserTheme} || $ConfigObject->Get('DefaultTheme') || Translatable('Standard');
+
+    # force a theme based on host name
+    my $DefaultThemeHostBased = $ConfigObject->Get('DefaultTheme::HostBased');
+    if ( $DefaultThemeHostBased && $ENV{HTTP_HOST} ) {
+
+        THEME:
+        for my $RegExp ( sort keys %{$DefaultThemeHostBased} ) {
+
+            # do not use empty regexp or theme directories
+            next THEME if !$RegExp;
+            next THEME if $RegExp eq '';
+            next THEME if !$DefaultThemeHostBased->{$RegExp};
+
+            # check if regexp is matching
+            if ( $ENV{HTTP_HOST} =~ /$RegExp/i ) {
+                $Theme = $DefaultThemeHostBased->{$RegExp};
+            }
+        }
+    }
+
+    # locate template files
+    my $JSStandardTemplateDir = $ConfigObject->Get('TemplateDir') . '/JavaScript/Templates/' . 'Standard';
+    my $JSTemplateDir         = $ConfigObject->Get('TemplateDir') . '/JavaScript/Templates/' . $Theme;
+
+    # Check if 'Standard' fallback exists
+    if ( !-e $JSStandardTemplateDir ) {
+        $Self->FatalDie(
+            Message =>
+                "No existing template directory found ('$JSTemplateDir')! Check your Home in Kernel/Config.pm."
+        );
+    }
+
+    if ( !-e $JSTemplateDir ) {
+        $Kernel::OM->Get('Kernel::System::Log')->Log(
+            Priority => 'error',
+            Message =>
+                "No existing template directory found ('$JSTemplateDir')!.
+                Default theme used instead.",
+        );
+
+        # Set TemplateDir to 'Standard' as a fallback.
+        $Theme         = 'Standard';
+        $JSTemplateDir = $JSStandardTemplateDir;
+    }
+
+    my $JSCustomStandardTemplateDir = $ConfigObject->Get('CustomTemplateDir') . '/JavaScript/Templates/' . 'Standard';
+    my $JSCustomTemplateDir         = $ConfigObject->Get('CustomTemplateDir') . '/JavaScript/Templates/' . $Theme;
+
+    my @TemplateFolders = (
+        "$JSCustomTemplateDir",
+        "$JSCustomStandardTemplateDir",
+        "$JSTemplateDir",
+        "$JSStandardTemplateDir",
+    );
+
+    my $JSHome               = $ConfigObject->Get('Home') . '/var/httpd/htdocs/js';
+    my $TargetFilenamePrefix = "TemplateJS";
+
+    my %ChecksumData;
+    my $TemplateChecksum;
+    my $MainObject = $Kernel::OM->Get('Kernel::System::Main');
+
+    TEMPLATEFOLDER:
+    for my $TemplateFolder (@TemplateFolders) {
+
+        next TEMPLATEFOLDER if !-e $TemplateFolder;
+
+        # get main object
+        my @Templates = $MainObject->DirectoryRead(
+            Directory => $TemplateFolder,
+            Filter    => '*.tmpl',
+            Recursive => 1,
+        );
+
+        TEMPLATE:
+        for my $Template ( sort @Templates ) {
+
+            next TEMPLATE if !-e $Template;
+
+            my $Key = $Template;
+            $Key =~ s/^$TemplateFolder\///xmsg;
+            $Key =~ s/\.(\w+)\.tmpl$//xmsg;
+
+            # check if a template with this name does already exist
+            next TEMPLATE if $ChecksumData{$Key};
+
+            # get file metadata
+            my $Stat = stat($Template);
+            if ( !$Stat ) {
+                print STDERR "Error: cannot stat file '$Template': $!";
+                next TEMPLATE;
+            }
+
+            $ChecksumData{$Key} = $Template . $Stat->mtime();
+        }
+    }
+
+    # generate a checksum only of the actual used files
+    for my $Checksum ( sort keys \%ChecksumData ) {
+        $TemplateChecksum .= $ChecksumData{$Checksum};
+    }
+    $TemplateChecksum = Digest::MD5::md5_hex($TemplateChecksum);
+
+    # Check if cache already exists.
+    if ( -e "$JSHome/js-cache/${TargetFilenamePrefix}_$TemplateChecksum.js" ) {
+        $Self->Block(
+            Name => 'CommonJS',
+            Data => {
+                JSDirectory => 'js-cache/',
+                Filename    => "${TargetFilenamePrefix}_$TemplateChecksum.js",
+            },
+        );
+        return 1;
+    }
+
+    # if it doesnt exist, go through the folders and get the template content
+    my %TemplateData;
+
+    TEMPLATEFOLDER:
+    for my $TemplateFolder (@TemplateFolders) {
+
+        next TEMPLATEFOLDER if !-e $TemplateFolder;
+
+        # get main object
+        my @Templates = $MainObject->DirectoryRead(
+            Directory => $TemplateFolder,
+            Filter    => '*.tmpl',
+            Recursive => 1,
+        );
+
+        TEMPLATE:
+        for my $Template ( sort @Templates ) {
+
+            next TEMPLATE if !-e $Template;
+
+            my $Key = $Template;
+            $Key =~ s/^$TemplateFolder\///xmsg;
+            $Key =~ s/\.(\w+)\.tmpl$//xmsg;
+
+            # check if a template with this name does already exist
+            next TEMPLATE if $TemplateData{$Key};
+
+            my $TemplateContent = ${
+                $MainObject->FileRead(
+                    Location => $Template,
+                    Result   => 'SCALAR',
+                )
+            };
+
+            # Remove DTL-style comments (lines starting with #)
+            $TemplateContent =~ s/^#.*\n//gm;
+            $TemplateData{$Key} = $TemplateContent;
+        }
+    }
+
+    my $TemplateDataJSON = $Kernel::OM->Get('Kernel::System::JSON')->Encode(
+        Data   => \%TemplateData,
+        Pretty => 0,
+    );
+
+    my $Content = <<"EOF";
+// The content of this file is automatically generated, do not edit.
+Core.Template.Load($TemplateDataJSON);
+EOF
+    my $MinifiedFile = $Kernel::OM->Get('Kernel::System::Loader')->MinifyFiles(
+        Checksum             => $TemplateChecksum,
+        Content              => $Content,
+        Type                 => 'JavaScript',
+        TargetDirectory      => "$JSHome/js-cache/",
+        TargetFilenamePrefix => $TargetFilenamePrefix,
+    );
+
+    $Self->Block(
+        Name => 'CommonJS',
+        Data => {
+            JSDirectory => 'js-cache/',
+            Filename    => $MinifiedFile,
+        },
+    );
 
     return 1;
 }